Move batchMode to ConsumerProperties.
