+++ spring-cloud-stream/src/main/java/org/springframework/cloud/stream/function/StreamFunctionProperties.java
@@ -40,8 +40,6 @@ public class StreamFunctionProperties {
 
 	private String outputDestinationName = Processor.OUTPUT;
 
-	private boolean batchMode;
-
 	public String getDefinition() {
 		return this.definition;
 	}
@@ -72,14 +70,6 @@ public class StreamFunctionProperties {
 
 	void setOutputDestinationName(String outputDestinationName) {
 		this.outputDestinationName = outputDestinationName;
-	}
-
-	public boolean isBatchMode() {
-		return this.batchMode;
-	}
-
-	public void setBatchMode(boolean batchMode) {
-		this.batchMode = batchMode;
 	}
 
 }
